Move code from apirule_controller to new files
